A lot of things added
So I took this time to add new things to the project, the dash movement is done, tried putting little twists, went back. Added hp to the player that is already synchronized to the UI, and a little crosshair along with something that tells the player when he can dash. Along with that, the initial level randomization teleport is done.