Rearchitecture of Systems
Rearchitecture of some variables that need to be fixed throught levels, for that I implemented the Game Instance class so that we can play with those.
fixed jump-dash bug, by making it so that the isdashing variable is disabled after .2 seconds, this is enough, maybe even too much lmfao
New input actions added, the combat inputs are already implemented, the implemenatation is a bit different to do, but it is here though.
Starting to improve the randomization of levels by mixing in the rewards too, for this we have the enum that has all the rewards the player can get, and then we'll mix and match them.